Consolidating src/ and include/ into  nematode/. Updating nmeaparse include to nematode include.
