+++ .gitignore
@@ -1,0 +1,4 @@
+CMakeLists.txt.user
+TriBITS/
+googletest/
+testframework/
