+++ PackagesList.cmake
@@ -1,0 +1,26 @@
+##---------------------------------------------------------------------------##
+## NemaTode/PackagesList.cmake
+## Jordan Lefebvre, lefebvrejp@ornl.gov
+##---------------------------------------------------------------------------##
+#
+# See documentation in TriBITS https://tribits.org
+
+
+##---------------------------------------------------------------------------##
+## PACKAGES PROVIDED
+##---------------------------------------------------------------------------##
+
+TRIBITS_REPOSITORY_DEFINE_PACKAGES(
+  googletest    googletest/googletest ST
+  testframework testframework         ST
+  nematode      .                     PT
+)
+
+TRIBITS_ALLOW_MISSING_EXTERNAL_PACKAGES(
+   googletest
+   testframework
+)
+
+##---------------------------------------------------------------------------##
+##                     end of NemaTode/PackagesList.cmake
+##---------------------------------------------------------------------------##
