+++ ProjectName.cmake
@@ -1,0 +1,12 @@
+##---------------------------------------------------------------------------##
+## NemaTode/ProjectName.cmake
+## Jordan Lefebvre, lefebvrejp@ornl.gov
+##---------------------------------------------------------------------------##
+
+# Must set the project name as a variable at very beginning before including
+# anything else
+SET(PROJECT_NAME NemaTode)
+
+##---------------------------------------------------------------------------##
+##                    end of NemaTode/ProjectName.cmake
+##---------------------------------------------------------------------------##
