+++ cmake/Dependencies.cmake
@@ -1,0 +1,10 @@
+TRIBITS_PACKAGE_DEFINE_DEPENDENCIES(
+LIB_REQUIRED_PACKAGES daemmodel waspjson waspcore radixgeo radixio radixcore
+LIB_OPTIONAL_PACKAGES
+TEST_REQUIRED_PACKAGES testframework
+TEST_OPTIONAL_PACKAGES
+LIB_REQUIRED_TPLS Qt5Widgets Qt5Location Qt5Positioning Qt5SerialPort Qt5Core
+LIB_OPTIONAL_TPLS
+TEST_REQUIRED_TPLS
+TEST_OPTIONAL_TPLS
+)
