+++ cmake/ExtraRepositoriesList.cmake
@@ -1,0 +1,5 @@
+TRIBITS_PROJECT_DEFINE_EXTRA_REPOSITORIES(
+  TriBITS  "" GIT https://github.com/lefebvre/TriBITS.git "NOPACKAGES" Continuous
+  TESTF_REPO  "testframework" GIT https://github.com/lefebvre/testframework "NOPACKAGES" Continuous
+  GTEST_REPO  "googletest" GIT https://github.com/lefebvre/googletest "NOPACKAGES" Continuous
+)
