+++ nematode/GPSFix.cpp
@@ -7,7 +7,7 @@
  *  See the license file included with this source.
  */
 
-#include <nmeaparse/GPSFix.h>
+#include <nematode/GPSFix.h>
 #include <cmath>
 #include <string>
 #include <sstream>
@@ -223,11 +223,11 @@ GPSFix::GPSFix() {
 
 	haslock = 0;
 
-	dilution = 0;		
+	dilution = 0;
 	horizontalDilution = 0;		// Horizontal - Best is 1, >20 is terrible, so 0 means uninitialized
-	verticalDilution = 0;	
-	latitude = 0;	
-	longitude = 0;	
+	verticalDilution = 0;
+	latitude = 0;
+	longitude = 0;
 	speed = 0;
 	travelAngle = 0;
 	altitude = 0;
@@ -324,7 +324,7 @@ std::string GPSFix::travelAngleToCompassDirection(double deg, bool abbrev){
 		};
 		return dirs[r];
 	}
-	
+
 };
