+++ nematode/NMEAParser.h
@@ -11,7 +11,7 @@
 #define NMEAPARSER_H_
 
 
-#include <nmeaparse/Event.h>
+#include <nematode/Event.h>
 #include <string>
 #include <functional>
 #include <unordered_map>
@@ -30,7 +30,7 @@
 
 namespace nmea {
 
-class NMEAParser; 
+class NMEAParser;
 
 
 class NMEASentence {
@@ -91,7 +91,7 @@ private:
 	uint32_t maxbuffersize;		//limit the max size if no newline ever comes... Prevents huge buffer string internally
 
 	void parseText	(NMEASentence& nmea, std::string s);		//fills the given NMEA sentence with the results of parsing the string.
-	
+
 	void onInfo		(NMEASentence& n, std::string s);
 	void onWarning	(NMEASentence& n, std::string s);
 	void onError	(NMEASentence& n, std::string s);
